Drop default version for App Engine Standard facet
When unspecified, WTP chooses the latest version of the facet as the default.
Our comparator ensures the JRE7 sorts before JRE8, and so the JRE8 will
appear as the default (if present).

In practice, this doesn't matter our New Project wizard prompts the user
for the runtime, and the default is set in AppEngineStandardWizardPage.