update readme with supported Eclipse versions
update readme with info to reflect that Eclipse 4.22 and later are currently not supported. Eclipse latest is currently on 4.24 now and 4.25 will come out soon.
Related issues got fired in #3659, #3660, #3678; attempts to fix in #3661 but no progress since.
Suggest adding this to readme and removing when this issue actually get fixed.